RsClModel/requirements.txt
absl-py==0.8.0
astor==0.8.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
gast==0.3.3
gitdb2==2.0.5
GitPython==2.1.11
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio
h5py==2.10.0
idna==2.8
importlib-metadata==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.1.1
numpy==1.16.4
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==0.25.0
Pillow==6.1.0
protobuf==3.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.2.4
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.1
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.12.0
smmap2==2.0.5
tabulate==0.8.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
torch==1.1.0
torchnet==0.0.4
torchvision==0.3.0
tornado==6.0.3
tqdm==4.35.0
urllib3==1.25.3
visdom==0.1.8.8
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.1
XlsxWriter==1.1.8
zipp==3.1.0
tensorboardX
geatpy